+++ ydb/core/kafka_proxy/actors/kafka_alter_configs_actor.cpp
@@ -64,7 +64,6 @@ public:
     ) {
         Y_UNUSED(selfInfo);
         const auto& pqConfig = appData->PQConfig;
-
         auto partitionConfig = groupConfig.MutablePQTabletConfig()->MutablePartitionConfig();
 
         if (RetentionMs.has_value()) {
@@ -75,7 +74,7 @@ public:
             partitionConfig->SetStorageLimitBytes(RetentionBytes.value());
         }
         if (CleanupPolicy.has_value()) {
-            tabletConfig->SetEnableCompactification(CleanupPolicy.value() == ECleanupPolicy::COMPACT);
+            groupConfig.MutablePQTabletConfig()->SetEnableCompactification(CleanupPolicy.value() == ECleanupPolicy::COMPACT);
         }
         if (pqGroupDescription.GetPQTabletConfig().GetEnableCompactification() && !groupConfig.GetPQTabletConfig().GetEnableCompactification()) {
             NKikimr::NGRpcProxy::V1::RemoveReadRuleFromConfig(
