+++ ydb/core/persqueue/partition.h
@@ -13,6 +13,7 @@
 #include "utils.h"
 #include "read_quoter.h"
 #include "partition_blob_encoder.h"
+#include "partition_compactification.h"
 
 #include <ydb/core/keyvalue/keyvalue_events.h>
 #include <ydb/library/persqueue/counter_time_keeper/counter_time_keeper.h>
@@ -363,7 +364,7 @@ private:
                         const TActorContext& ctx);
 
 
-    void ScheduleReplyOk(const ui64 dst);
+    void ScheduleReplyOk(const ui64 dst, bool internal = false);
     void ScheduleReplyGetClientOffsetOk(const ui64 dst,
                                         const i64 offset,
                                         const TInstant writeTimestamp,
@@ -1091,98 +1092,6 @@ private:
     TInstant GetFirstUncompactedBlobTimestamp() const;
 
     void TryCorrectStartOffset(TMaybe<ui64> offset);
-};
-
-
-class TPartitionCompaction {
-public:
-    TPartitionCompaction(ui64 lastCompactedOffset, ui64 startCookie, ui64 endCookie, TPartition* partitionActor);
-
-    enum class EStep {
-        PENDING,
-        READING,
-        COMPACTING,
-    };
-
-    class TReadState {
-        friend TPartitionCompaction;
-        constexpr static const ui64 MAX_DATA_KEYS = 5000;
-
-        ui64 OffsetToRead;
-        ui64 LastOffset = 0;
-        ui64 NextPartNo = 0;
-        THashMap<TString, ui64> TopicData; //Key -> Offset
-
-
-        TPartition* PartitionActor;
-        TMaybe<NKikimrClient::TCmdReadResult::TResult> LastMessage;
-
-    public:
-        TReadState(ui64 firstOffset, TPartition* partitionActor);
-
-        bool ProcessResponse(TEvPQ::TEvProxyResponse::TPtr& ev);
-        EStep ContinueIfPossible(ui64 nextRequestCookie);
-        THashMap<TString, ui64>&& GetData();
-        ui64 GetLastOffset();
-        void UpdateConfig(ui64 maxBurst, ui64 readQuota); //ToDo;
-    };
-
-    class TCompactState {
-        friend TPartitionCompaction;
-        using TKeysIter = std::deque<TDataKey>::iterator;
-
-        THolder<TEvKeyValue::TEvRequest> Request;
-        ui64 RequestSize;
-        ui64 MaxOffset;
-        THashMap<TString, ui64> TopicData;
-        TVector<TKey> DroppedKeys;
-        TPartition* PartitionActor;
-
-        TMaybe<ui64> DropOffset;
-        TKeysIter KeysIter;
-        TSet<ui64> OffsetsToKeep;
-        TSet<ui64>::iterator OffsetsIter;
-        bool Failure = false;
-        ui64 RequestDataSize = 0;
-        std::deque<TDataKey> DataKeysBody;
-        ui64 FirstHeadOffset;
-        ui64 FirstHeadPartNo;
-
-        TCompactState(THashMap<TString, ui64>&& data, ui64 firstUncompactedOffset, ui64 maxOffset, TPartition* partitionActor);
-
-        EStep ProcessKVResponse(TEvKeyValue::TEvResponse::TPtr& ev);
-        bool ProcessResponse(TEvPQ::TEvProxyResponse::TPtr& ev);
-
-        EStep ContinueIfPossible(ui64 nextCookie);
-        void RunKvRequest();
-        void AddDeleteRange(const TKey& key);
-    };
-
-    EStep Step = EStep::PENDING;
-    ui64 FirstUncompactedOffset = 0;
-    bool HaveRequestInflight = false;
-    ui64 StartCookie = 0;
-    ui64 EndCookie = 0;
-    ui64 CurrentCookie = 0;
-
-
-    std::optional<TReadState> ReadState;
-    std::optional<TCompactState> CompactState;
-    TPartition* PartitionActor;
-
-    ui64 GetNextCookie() {
-        CurrentCookie++;
-        if (CurrentCookie > EndCookie) {
-            CurrentCookie = StartCookie;
-        }
-        return CurrentCookie;
-    }
-
-    void TryCompactionIfPossible();
-    void UpdatePartitionConifg();
-    void ProcessResponse(TEvPQ::TEvProxyResponse::TPtr& ev);
-    void ProcessResponse(TEvKeyValue::TEvResponse::TPtr& ev);
-
 };
 
 } // namespace NKikimr::NPQ
