+++ ydb/core/persqueue/ut/pq_ut.cpp
@@ -1443,6 +1443,7 @@ Y_UNIT_TEST(TestComactifiedWithRetention) {
         tc.Prepare(dispatchName, setup, activeZone);
 
         tc.Runtime->SetScheduledLimit(100);
+        tc.Runtime->GetAppData(0).PQConfig.MutableCompactionConfig()->SetBlobsCount(0);
 
         TVector<std::pair<ui64, TString>> data;
         activeZone = PlainOrSoSlow(true, false);
@@ -1457,10 +1458,12 @@ Y_UNIT_TEST(TestComactifiedWithRetention) {
         CmdWrite(0, "sourceid1", data, tc, false);
         CmdWrite(0, "sourceid2", data, tc, false);
         PQGetPartInfo(0, 30, tc);
+
         PQTabletPrepare({.maxCountInPartition=1000, .deleteTime=0, .lowWatermark=100, .enableCompactificationByKey = false}, {}, tc);
         CmdWrite(0, "sourceid3", data, tc, false);
         CmdWrite(0, "sourceid4", data, tc, false);
         CmdWrite(0, "sourceid5", data, tc, false);
+        Cerr << "Get part info with compactification disabled\n";
         PQGetPartInfo(50, 60, tc);
     });
 }
